+++ client/src/VespaClusterInfo.ts
@@ -1,7 +1,7 @@
 import exp = require('constants');
 import { vespaConfig } from './VespaConfig';
 import { outputChannel, showError } from './extension';
-import { Schema, SchemaConfig } from './model/VespaSchemaConfig';
+import { VespaSchema, VespaSchemaConfig } from './model/VespaSchemaConfig';
 import { StorageClusterConfig, VespaClusterList } from './model/VespaClusterList';
 import { VespaAppId } from './model/VespaAppId';
 import { fetchWithTimeout } from './vespaUtils';
@@ -12,13 +12,13 @@ export class VespaClusterInfo {
 	clusterName: string = undefined;
 	appId: VespaAppId = undefined;
 	clusterList: VespaClusterList = undefined;
-	clusterSchemas: Map<string, Schema[]> = new Map();
+	clusterSchemas: Map<string, VespaSchema[]> = new Map();
 
 	defaultConfigId() {
 		return this.clusterList.storage[0].configid;
 	}
 
-	defaultClusterSchemas(): Schema[] {
+	defaultClusterSchemas(): VespaSchema[] {
 		return this.clusterSchemas.get(this.defaultConfigId());
 	}
 
@@ -31,11 +31,10 @@ export class VespaClusterInfo {
 		return fetchWithTimeout(clusterListUrl, timeout)
 			.then(response => response.json()
 				.then(v => v as VespaClusterList));
-
 	}
 
 
-	fetchSchemas(storageConfig: StorageClusterConfig): Promise<Schema[]> {
+	fetchSchemas(storageConfig: StorageClusterConfig): Promise<VespaSchema[]> {
 		const configEndpoint: string = vespaConfig.configEndpoint();
 		const configid = storageConfig.configid;
 		const schemaListUrl = `${configEndpoint}/config/v2/tenant/${this.appId.tenant}/application/${this.appId.application}/search.config.schema-info/${configid}/search/cluster.${configid}`;
@@ -44,7 +43,7 @@ export class VespaClusterInfo {
 		outputChannel.appendLine(`Getting Schemas for ${configid} from ${configEndpoint}`);
 		return fetchWithTimeout(schemaListUrl, timeout)
 			.then(response => response.json()
-				.then(v => v as SchemaConfig))
+				.then(v => v as VespaSchemaConfig))
 			.then(schemaConfig => {
 				return schemaConfig.schema;
 			});
@@ -61,7 +60,7 @@ export class VespaClusterInfo {
 						.then(clusterList => {
 							clusterList.storage.map(storageConfig => {
 								return this.fetchSchemas(storageConfig)
-									.then((schemas: Schema[]) => {
+									.then((schemas: VespaSchema[]) => {
 										this.clusterSchemas.set(storageConfig.configid, schemas);
 									})
 									.catch(error => showError(error.reason));