+++ client/src/VespaConfigExplorer.ts
@@ -1,13 +1,15 @@
 import * as vscode from 'vscode';
 import { outputChannel, showError } from './extension';
-import { homedir } from 'os';
 import fs = require('fs');
 import * as path from 'path';
-import { EventEmitter } from 'stream';
 import { vespaConfig } from './VespaConfig';
+import { VespaSchemaField, VespaSchema, VespaSchemaConfig } from './model/VespaSchemaConfig';
+import { VespaAppId } from './model/VespaAppId';
+import { VespaClusterList } from './model/VespaClusterList';
 import { VespaStatusResultsPanel } from './VespaStatusResultsPanel';
+import { VespaV2Metrics } from './model/VespaMetrics';
 
-export class VespaConfigView {
+export class VespaConfigExplorer {
 
 	vespaConfigProvider: VespaConfigProvider;
 
@@ -18,6 +20,7 @@ export class VespaConfigView {
 		this.vespaConfigProvider = new VespaConfigProvider(context);
 
 		this.registerSetDefaultCommand();
+		this.registerShowClusterStatusCommand(context);
 
 		vscode.commands.registerCommand("vscode-vespa.refreshVespaConfig", args => {
 			if (args !== undefined) {
@@ -60,9 +63,34 @@ export class VespaConfigView {
 	}
 
 
-	
 
-	private registerSetDefaultCommand() {		
+	static showClusterStatus(extensionUri: vscode.Uri, clusterName: string) {
+
+		// outputChannel.appendLine("clusterName: " + clusterName);
+
+		const cluster = vespaConfig.getCluster(clusterName);
+		VespaV2Metrics.fetchDocInfo(cluster.configEndpoint)
+			.then(docInfo => {
+				VespaStatusResultsPanel.createOrShow(extensionUri, cluster, docInfo, new Date());
+			}).catch(error => {
+				showError("docCounts failed " + error + "\n" + error.stack);
+			});
+	}
+
+	private registerShowClusterStatusCommand(context: vscode.ExtensionContext) {
+		vscode.commands.registerCommand("vscode-vespa.showSelectedClusterInfo", args => {
+			if (args !== undefined) {
+				const { key } = args;
+				if (key.startsWith("cluster:")) {
+					const clusterName = key.replace("cluster:", "");
+					VespaConfigExplorer.showClusterStatus(context.extensionUri, clusterName);
+				}
+			}
+		});
+	}
+
+
+	private registerSetDefaultCommand() {
 		vscode.commands.registerCommand("vscode-vespa.selectClusterAsDefault", args => {
 			if (args !== undefined) {
 				const { key } = args;
@@ -131,10 +159,10 @@ const defaultClusterItem = new vscode.TreeItem("defaultCluster");
 const clustersItem = new vscode.TreeItem("clusters");
 
 
-export class VespaConfigProvider implements vscode.TreeDataProvider<{ key: string }> {
+export class VespaConfigProvider implements vscode.TreeDataProvider<{ key: string, data: any }> {
 
-	private _onDidChangeTreeData: vscode.EventEmitter<{ key: string } | undefined> = new vscode.EventEmitter<{ key: string } | undefined>();
-	readonly onDidChangeTreeData: vscode.Event<{ key: string } | undefined> = this._onDidChangeTreeData.event;
+	private _onDidChangeTreeData: vscode.EventEmitter<{ key: string, data: any } | undefined> = new vscode.EventEmitter<{ key: string, data: any } | undefined>();
+	readonly onDidChangeTreeData: vscode.Event<{ key: string, data: any } | undefined> = this._onDidChangeTreeData.event;
 
 	private context: vscode.ExtensionContext;
 
@@ -146,40 +174,104 @@ export class VespaConfigProvider implements vscode.TreeDataProvider<{ key: strin
 		this._onDidChangeTreeData.fire(undefined);
 	}
 
-	getChildren(element: { key: string }): { key: string }[] {
+	getChildren(element: { key: string, data: any }): Thenable<{ key: string, data: any }[]> {
 
 		if (element !== undefined && element.key.startsWith("cluster:")) {
 
 			const clusterName = element.key.replace("cluster:", "");
 			// outputChannel.appendLine("clusterName: " +clusterName + ", default: " + vespaConfig.defaultCluster);
-			return [
-				{ key: `queryEndpoint|${clusterName}` },
-				{ key: `configEndpoint|${clusterName}` },
-				{ key: `zipkinEndpoint|${clusterName}` },
-			];
+			return Promise.resolve([
+				//{ key: `queryEndpoint|${clusterName}`, data: clusterName, },
+				//{ key: `configEndpoint|${clusterName}`, data: clusterName, },
+				//{ key: `zipkinEndpoint|${clusterName}`, data: clusterName, },
+				{ key: `docTypes|${clusterName}`, data: clusterName, },
+			]);
 		}
 
+		if (element !== undefined && element.key.startsWith("docTypes|")) {
+			const clusterName = element.key.replace("docTypes|", "");
+			outputChannel.appendLine("DOC types for: " + clusterName);
+
+			const cluster = vespaConfig.getCluster(clusterName);
+			if (cluster) {
+				const configEndpoint = cluster.configEndpoint;
+				return VespaAppId.fetchAppId(configEndpoint)
+					.then(appiId => {
+						return VespaClusterList.fetchVespaClusterList(configEndpoint, appiId)
+							.then(clusterList => {
+								return VespaSchemaConfig.fetchSchemas(configEndpoint, appiId, clusterList.getDefaultConfigId())
+									.then(schemaConfig => {
+										return schemaConfig.schema.map(schema => {
+											outputChannel.appendLine("schema: " + schema.name);
+											return {
+												key: `docType|${schema.name}`,
+												data: {
+													clusterName: clusterName,
+													configId: clusterList.getDefaultConfigId(),
+													schema: schema
+												},
+											};
+										});
+									});
+							});
+					});
+			} else {
+				return Promise.resolve([]);
+			}
+		}
+
+		if (element !== undefined && element.key.startsWith("docType|")) {
+			const docType = element.key.replace("docType|", "");
+			outputChannel.appendLine("Fields for: " + docType);
+
+			const data = element.data as {
+				clusterName: string;
+				configId: string;
+				schema: VespaSchema;
+			};
+
+			if (data) {
+				// const configEndpoint = vespaConfig.getCluster(data.clusterName).configEndpoint;
+				return Promise.resolve(data.schema.field.map(field => {
+					return {
+						key: "field",
+						data: {
+							field: field,
+						}
+					};
+				}));
+			} else {
+				return Promise.resolve([]);
+			}
+		}
 
 		if (element !== undefined) {
 			// return getChildren(element ? element.key : undefined).map(key => getNode(key));
 			switch (element.key) {
 				case "vespa-config":
-					return [
-						{ key: "defaultCluster" },
-						{ key: "clusters" },
-					];
+					return Promise.resolve([
+						{ key: "defaultCluster", data: undefined, },
+						{ key: "clusters", data: undefined, },
+					]);
 				case "clusters":
-					return vespaConfig.clusters().map(cluster => {
-						return { key: `cluster:${cluster.name}` };
-					});
+					return Promise.resolve(vespaConfig.clusters().map(cluster => {
+						return { key: `cluster:${cluster.name}`, data: cluster, };
+					}));
 			}
 		}
-		return [
-			{ key: "vespa-config" },
-		];
+		if (element === undefined) {
+			// return Promise.resolve([
+			// 	{ key: "vespa-config", data: undefined, },
+			// ]);
+			return Promise.resolve([
+				{ key: "defaultCluster", data: undefined, },
+				{ key: "clusters", data: undefined, },
+			]);
+		}
+		return Promise.resolve([]);
 	}
 
-	getTreeItem(element: { key: string }): vscode.TreeItem {
+	getTreeItem(element: { key: string, data: any }): vscode.TreeItem {
 		// outputChannel.appendLine("getTreeItem(): " + JSON.stringify(element));
 
 		// {key: `queryEndpoint|${cluster.name}|${cluster.queryEndpoint}`},
@@ -216,6 +308,44 @@ export class VespaConfigProvider implements vscode.TreeDataProvider<{ key: strin
 		// {key: `configEndpoint|${cluster.name}|${cluster.configEndpoint}`},
 		// {key: `zipkinEndpoint|${cluster.name}|${cluster.zipkinEndpoint}`},
 
+		if (element.key.startsWith("docTypes|")) {
+			const keyParts = element.key.split("|");
+			const clusterName = keyParts[1];
+
+			return {
+				label: `doc-types`,
+				iconPath: new vscode.ThemeIcon('database'),
+				// tooltip: new vscode.MarkdownString(`$(zap) Tooltip for ${element.key}`, true),
+				collapsibleState: vscode.TreeItemCollapsibleState.Collapsed,
+				contextValue: "VESPA_EXPLORER_DOC_TYPES",
+			};
+		}
+
+		if (element.key.startsWith("docType|")) {
+			const keyParts = element.key.split("|");
+			const docType = keyParts[1];
+
+			return {
+				label: docType,
+				iconPath: new vscode.ThemeIcon('file'),
+				// tooltip: new vscode.MarkdownString(`$(zap) Tooltip for ${element.key}`, true),
+				collapsibleState: vscode.TreeItemCollapsibleState.Collapsed,
+				// contextValue: "VESPA_EXPLORER_DOC_TYPE",
+			};
+		}
+
+		if (element.key === "field") {
+			const data = element.data as {
+				field: VespaSchemaField;
+			};
+			return {
+				label: `${data.field.name}: ${data.field.type}`,
+				iconPath: new vscode.ThemeIcon('symbol-property'),
+				// tooltip: new vscode.MarkdownString(`$(zap) Tooltip for ${element.key}`, true),
+				collapsibleState: vscode.TreeItemCollapsibleState.None,
+				// contextValue: "VESPA_EXPLORER_DOC_TYPE",
+			};
+		}
 
 		if (element.key.startsWith("cluster:")) {
 
@@ -225,10 +355,13 @@ export class VespaConfigProvider implements vscode.TreeDataProvider<{ key: strin
 			return {
 				label: clusterName,
 				// iconPath: clusterName === vespaConfig.defaultCluster ?new vscode.ThemeIcon("$(zap)"): "",
-				iconPath: clusterName === vespaConfig.default() ? {
-					light: this.context.asAbsolutePath(path.join('media', 'boolean.svg')),
-					dark: this.context.asAbsolutePath(path.join('media', 'boolean.svg'))
-				} : "",
+				// iconPath: clusterName === vespaConfig.default() ? {
+				// 	light: this.context.asAbsolutePath(path.join('media', 'boolean.svg')),
+				// 	dark: this.context.asAbsolutePath(path.join('media', 'boolean.svg'))
+				// } : "",	
+				iconPath: clusterName === vespaConfig.default() ?
+					new vscode.ThemeIcon('check', new vscode.ThemeColor("testing.runAction"))
+					: "",
 				tooltip: new vscode.MarkdownString(`$(zap) Tooltip for ${element.key}`, true),
 				collapsibleState: vscode.TreeItemCollapsibleState.Collapsed,
 				contextValue: "VESPA_CLUSTER",
@@ -243,7 +376,7 @@ export class VespaConfigProvider implements vscode.TreeDataProvider<{ key: strin
 				};
 			case "defaultCluster":
 				return {
-					label: `defaultCluster: ${vespaConfig.default()}`,
+					label: `connected to: ${vespaConfig.default()}`,
 					//tooltip: new vscode.MarkdownString(`$(zap) Tooltip for ${element.key}`, true),
 				};
 			case "clusters":
@@ -256,7 +389,7 @@ export class VespaConfigProvider implements vscode.TreeDataProvider<{ key: strin
 		return {};
 	}
 
-	getParent?(element: { key: string }): { key: string } | undefined {
+	getParent?(element: { key: string, data: any }): { key: string, data: any } | undefined {
 		// outputChannel.appendLine("getParent(): " + element);
 
 		if (element.key.startsWith("queryEndpoint|") ||
@@ -268,7 +401,8 @@ export class VespaConfigProvider implements vscode.TreeDataProvider<{ key: strin
 			const clusterName = keyParts[1];
 			// const cluster = vespaConfig.getCluster(clusterName);
 			return {
-				key: `cluster:${clusterName}`
+				key: `cluster:${clusterName}`,
+				data: undefined,
 			};
 		}
 
@@ -277,6 +411,7 @@ export class VespaConfigProvider implements vscode.TreeDataProvider<{ key: strin
 			case "defaultCluster":
 				return {
 					key: "vespa-config",
+					data: undefined,
 				};
 		}
 		return undefined;