+++ CHANGELOG.md
@@ -2,9 +2,18 @@
 
 All notable changes to the "vscode-vespa" extension will be documented in this file.
 
-## [1.0.2]
+## [1.1.0]
 
-- Add button to save result json as file.
+- Add button to save YQL result json as file.
+- Support for multiple vespa clusters via custom config 
+- Move configuration to $HOME/.config/vscode-vespa/vespa-config.json
+- Vespa explorer to view details about the configured vespa clusters.
+- Vespa Cluster Info panel with document-info and iframe for Controller Status page.
+- "Vespa YQL Results" tab added to YQL results containing details about the query request.
+- editor/title menus for
+  - Connecting to cluster
+  - Edit Cluster Config 
+  - Show Vespa Cluster Info
 
 
 ## [1.0.1]